[#auvvph] - Dont allow to delete last user with sec_admin

+++ src/app/user-management/user-management.component.ts
@@ -6,6 +6,8 @@ import { AuthenticationService } from '../services/authentication.service';
 import { GroupDto } from '../services/dto/group.dto';
 import { NzMessageService } from 'ng-zorro-antd/message';
 
+export const SEC_ADMIN_GROUP_NAME: string = 'SEC_ADMIN';
+
 @Component({
   selector: 'app-user-management',
   templateUrl: './user-management.component.html',
@@ -13,7 +15,7 @@ import { NzMessageService } from 'ng-zorro-antd/message';
 })
 export class UserManagementComponent implements OnInit {
 
-  dataset: User[];
+  dataset: any[];
   visible: boolean = false;
   newUserForm: FormGroup;
   groupOptions: GroupDto[];
@@ -67,10 +69,14 @@ export class UserManagementComponent implements OnInit {
     this.selectedGroups = groups;
   }
 
-  deleteUser(user: User) {
-    this.userService.deleteUserByUsername(user.username).subscribe(res=> {
-      this.refreshUserTable();
-    });
+  deleteUser(user) {
+    if(user.groups.includes(SEC_ADMIN_GROUP_NAME) && this.dataset.filter(x=>x.groups.includes(SEC_ADMIN_GROUP_NAME)).length == 1){
+      this.messageService.error('Cannot delete only user with SEC_ADMIN group since then the application won\'t be accessible')
+    } else {
+      this.userService.deleteUserByUsername(user.username).subscribe(res=> {
+        this.refreshUserTable();
+      });
+    }
   }
 
   passwordsMatch: ValidatorFn = (control: FormGroup): ValidationErrors | undefined => {
